# src/main/resources/api/swagger.yml
swagger: '2.0'

info:
  title: Plan B Revocation API
  description: API for managing OAuth 2 token revocations
  version: "1.0"

securityDefinitions:
  oauth2:
    type: oauth2
    flow: implicit
    authorizationUrl: none
    scopes:
      uid: user identitiy

security:
  - oauth2:
    - uid

paths:
  /revocations:
    get:
      summary: Returns all the revocations since the specified timestamp.
      description: |
        Returns all revoked tokens from a specified timestamp. Also returns meta information which may be of importance to the client, like a refresh notification.
      tags:
        - Revocation
      parameters:
        - name: from
          description: instant from when tokens were revoked, in UTC UNIX timestamp
          in: query
          type: integer
          required: true
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/RevocationInfo'
        400:
          description: Type mismatch. When the parameter 'from' is missing or not an integer
          schema:
            $ref: '#/definitions/Problem'
    post:
      summary: Posts the specified revocation to be stored.
      description: |
        Revokes tokens associated with the specified revocation type.
      tags:
        - Revocation
      parameters:
        - name: revocation
          description: |
            the revocation associated with the tokens to revoke
          in: body
          schema:
            $ref: '#/definitions/Revocation'
      responses:
        # Response code
        201:
          description: Revocation was successfully stored
        400:
          description: body is not readable because it is not valid JSON, or have an invalid JSON structure
          schema:
            $ref: '#/definitions/Problem'
        500:
          description: an internal error occurred, such as a failure to contact the revocation store.
          schema:
            $ref: '#/definitions/Problem'

  /notifications/{type}:
    post:
      summary: Posts the specified notification to be stored.
      description: |
        Resource to post notification information, like a revocation refresh.
      tags:
        - Notification
      parameters:
        - name: type
          in: path
          description: |
            the type of notification
          required: true
          type: string
          enum: [ "REFRESH_FROM" ]
        - name: value
          in: query
          description: |
            the value of the notification to be stored
          required: true
          type: string
      responses:
        # Response code
        201:
          description: Notification was successfully stored
        400:
          description: type mismatch in parameter 'value'
          schema:
            $ref: '#/definitions/Problem'
        500:
          description: an internal error occurred, such as a failure to contact the revocation store.
          schema:
            $ref: '#/definitions/Problem'

definitions:
  RevocationInfo:
    type: object
    properties:
      meta:
        type: object
        properties:
          MAX_TIME_DELTA:
            type: integer
            description: Used to inform clients of the maximum time in seconds since when it is possible to get revocations.
          REFRESH_FROM:
            type: integer
            description: UNIX Timestamp (UTC) informing clients that a refresh of revocations needs to be done, starting from a specified instant.
          REFRESH_TIMESTAMP:
            type: integer
            description: UNIX Timestamp (UTC) informing clients when REFRESH_FROM was set.
      revocations:
        type: array
        items:
          $ref: '#/definitions/Revocation'
  Revocation:
    type: object
    properties:
      type:
        type: string
        enum: [ "TOKEN", "CLAIM", "GLOBAL" ]
      revoked_at:
        type: integer
      data:
        description: |
        type: object
        format: '#/definitions/Data'
  Data:
    type: object
  Claim:
    allOf:
    - $ref: '#/definitions/Data'
    - type: object
      properties:
        name:
          type: string
          description: The claim of the tokens to revoke
        value_hash:
          type: string
          description: The revoked claim hashed using hash_algorithm, in URL Base64 encoding
        hash_algorithm:
          type: string
          description: The algorithm used for hashing the Claim
        issued_before:
          type: integer
          description: a UNIX Timestamp (UTC) indicating that tokens issued before it are revoked. If this property is not set when posting a Claim Revocation, it will default to the current UNIX timestamp (UTC).
  Global:
    allOf:
    - $ref: '#/definitions/Data'
    - type: object
      properties:
        issued_before:
          type: integer
          description: a UNIX Timestamp (UTC) indication that tokens issued before it are revoked.
  Token:
    allOf:
    - $ref: '#/definitions/Data'
    - type: object
      properties:
        token_hash:
          type: string
          description: The revoked token hashed using hash_algorithm, in URL Base64 encoding
        hash_algorithm:
          type: string
          description: The algorithm used for hashing the Token
  Problem:
    type: object
    description: |
      Represents a problem as specified by the IETF Problem Details draft (03). Used to complement HTTP Status codes with additional contextual information.
    properties:
      type:
        type: string
        description: |
          URI set to "about:blank"
      title:
        type: string
        description: |
          reason for the specified status
      status:
        type: string
        description: |
          HTTP status code
      detail:
        type: string
        description: |
          additional detail of the problem